Remove the manual ContentLength calculation
Are you KITTEN me?

I was receiving this error when trying to push compressed files onto S3
while using this library.
`BadDigest: The Content-MD5 you specified did not match what we received.`

This library was automatically setting the length of the content to the
buffer.length value with; `s3Params.ContentLength = buffer.length;` for
the uploadBuffer call.

BUT, if the content is compressed, we aren't using regular ASCII
character, we're using the full UTF-8 range of characters.  That meant
that weird characters were worth more than one regular character, which
threw a BadDigest error.

BUT what's even more surprising is that UTF-8 is not the default
encoding for pushing content to S3 with this library.  After digging
into the code, I've came to realize that S3 automatically sets the
ContentLength of the buffer that this library was manually calculating.

We could either, force the encoding to UTF-8 OR let S3 handle that type
of stuff by removing the ContentLength calculation.

I voted for the laziest solution which was to remove a line of code.

And Guess what? We can now publish compressed data to S3 with this
library without errors.

The more you know! :)